Apply `LangVersion` `latest` and split into separate files
